Bring consistency among sizes
There are instances in the code that refer to lengths or sizes of strings and
buffers and the like, but they are not typed in a consistent manner. The purpose
of this commit is to provide some consistency to these scenarios by changing
their type from unsigned int, int, unsigned long, etc. to size_t.

In general, these situations were identified by globally searching the code
for patterns along the lines of:
(unsigned|unsigned int|int|long) (len|blen|olen|length|l|size)
and changing the type to size_t on a case-by-case basis.

The following situations were explicitly ignored:

- autosetup/jimsh0.c, because it appears to be third-party code

- inputs to function calls relating to ssl and sasl, because they don't take
size_t as inputs and recasting in the function call didn't seem worth the
increased messiness

- for loop indices, because in some cases unsigned indices can breed bugs

This commit takes care of most obvious cases. There may certainly be subtler
ones that were missed in this pass.

Resolves: #928

+++ attach.c
@@ -307,7 +307,7 @@ bailout:
  * @param type Buffer with mime type of attachment in "type/subtype" format
  * @param len  Buffer length
  */
-void mutt_check_lookup_list(struct Body *b, char *type, int len)
+void mutt_check_lookup_list(struct Body *b, char *type, size_t len)
 {
   int i;
